Add files via upload
Modified traffic data JSON files to change "Republic of Korea" to "South Korea" under the countryLbl array.

Added instrumentData3.json

Added two lines at the end of countries.csv in case the dataset includes strings "undefined" or "null" in the countryLbl array. These might be unused, but it may be better to have them. This will set lat=0 and long=0 to that particular data point to avoid further bugs in drawing the map.